util: move labels.go and objectmeta.go to util
