+++ controllers/util/labels.go
@@ -14,21 +14,21 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controllers
+package util
 
 const (
 	labelOwnerNamespaceName = "ramendr.openshift.io/owner-namespace-name"
 	labelOwnerName          = "ramendr.openshift.io/owner-name"
 )
 
-func ownerLabels(ownerNamespaceName, ownerName string) map[string]string {
+func OwnerLabels(ownerNamespaceName, ownerName string) map[string]string {
 	return map[string]string{
 		labelOwnerNamespaceName: ownerNamespaceName,
 		labelOwnerName:          ownerName,
 	}
 }
 
-func ownerNamespaceNameAndName(labels map[string]string) (string, string, bool) {
+func OwnerNamespaceNameAndName(labels map[string]string) (string, string, bool) {
 	ownerNamespaceName, ok1 := labels[labelOwnerNamespaceName]
 	ownerName, ok2 := labels[labelOwnerName]
